+++ monitor/structures.c
@@ -46,7 +46,7 @@ void test_struct(struct foo *x) {
 
 char serial_getchar(unsigned short port) {
   unsigned short result;
-  volatile unsigned short *p;
+  volatile unsigned int *p;
 
   switch (port) {
   case 0:
@@ -77,7 +77,7 @@ char *itos(unsigned int val, char *s) {
 }
   
 void serial_putchar(unsigned short port, char c) {
-  volatile unsigned short *p;
+  volatile unsigned int *p;
 
   switch (port) {
   case 0:
